+++ raknet/RakNetSocket2_Berkley.cpp
@@ -479,7 +479,7 @@ void RNS2_Berkley::RecvFromBlockingIPV4(RNS2RecvStruct *recvFromStruct)
 		sockAddrPtr=(sockaddr*) &sa;
 	}
 
-	recvFromStruct->bytesRead = recvfrom__( GetSocket(), recvFromStruct->data, sizeof(recvFromStruct->data), flag, sockAddrPtr, socketlenPtr );
+	recvFromStruct->bytesRead = (int)recvfrom__( GetSocket(), recvFromStruct->data, sizeof(recvFromStruct->data), flag, sockAddrPtr, socketlenPtr );
